+++ tests/cmds/test_high_risk_employee.py
@@ -405,8 +405,8 @@ def test_bulk_remove_risk_tags_uses_expected_arguments(runner, cli_state, mocker
 def test_remove_high_risk_employee_when_user_not_on_list_prints_expected_error(
     mocker, runner, cli_state
 ):
-    cli_state.sdk.detectionlists.high_risk_employee.remove.side_effect = get_user_not_on_list_side_effect(
-        mocker, "high-risk-employee"
+    cli_state.sdk.detectionlists.high_risk_employee.remove.side_effect = (
+        get_user_not_on_list_side_effect(mocker, "high-risk-employee")
     )
     test_username = "test@example.com"
     result = runner.invoke(